common_recordbatch

Function map_json_type_to_string

pub fn map_json_type_to_string(
    batch: RecordBatch,
    original_schema: &SchemaRef,
    mapped_schema: &SchemaRef,
) -> Result<RecordBatch>

Maps the json type to string in the batch.

The json type is mapped to string by converting the json value to string. The batch is updated to have the same number of columns as the original batch, but with the json type mapped to string.