common_recordbatch

Module util

Structs

ChainedRecordBatchStream

A stream that chains multiple streams into a single stream.

Functions

collect

Collect all the items from the stream into a vector of RecordBatch.

collect_batches

Collect all the items from the stream into RecordBatches.