Crate common_recordbatch

Re-exports

pub use recordbatch::RecordBatch;

Modules

adapter

cursor

error

Error of record batch.

filter

Util record batch stream wrapper that can perform precise filter.

recordbatch

util

Structs

DfRecordBatch

A two-dimensional batch of column-oriented data with a defined schema.

EmptyRecordBatchStream

EmptyRecordBatchStream can be used to create a RecordBatchStream that will produce no results

MemoryPermit

Memory permit for a stream, providing privileged access or rate limiting.

MemoryTrackedStream

A wrapper stream that tracks memory usage of RecordBatches.

OrderOption

QueryMemoryTracker

Memory tracker for RecordBatch streams. Clone to share the same limit across queries.

RecordBatchStreamWrapper

Adapt a [Stream] of RecordBatch to a RecordBatchStream.

RecordBatches

SendableRecordBatchMapper

A wrapper that maps a RecordBatchStream to a new RecordBatchStream by applying a function to each RecordBatch.

SimpleRecordBatchStream

Traits

RecordBatchStream

Functions

map_json_type_to_string

Maps the json type to string in the batch.

map_json_type_to_string_schema

Maps the json type to string in the schema.

Type Aliases

DfSendableRecordBatchStream

Trait for a [Stream] of RecordBatches that can be passed between threads

SendableRecordBatchStream