Crate common_recordbatch

Modules

adapter

cursor

error

Error of record batch.

filter

Util record batch stream wrapper that can perform precise filter.

recordbatch ๐

util

Structs

DfRecordBatch

A two-dimensional batch of column-oriented data with a defined schema.

EmptyRecordBatchStream

EmptyRecordBatchStream can be used to create a RecordBatchStream that will produce no results

OrderOption

RecordBatch

A two-dimensional batch of column-oriented data with a defined schema.

RecordBatchStreamWrapper

Adapt a [Stream] of RecordBatch to a RecordBatchStream.

RecordBatches

SendableRecordBatchMapper

A wrapper that maps a RecordBatchStream to a new RecordBatchStream by applying a function to each RecordBatch.

SimpleRecordBatchStream

Traits

RecordBatchStream

Functions

map_json_type_to_string

Maps the json type to string in the batch.

map_json_type_to_string_schema

Maps the json type to string in the schema.

Type Aliases

DfSendableRecordBatchStream

Trait for a [Stream] of RecordBatches that can be passed between threads

SendableRecordBatchStream